Forward the query context and filter the projection by account
The wrapper called the SDK query function with no arguments, so the
abort signal React Query hands it never reached the bridge call: a
wallet the reader had navigated away from kept downloading and
resolving entries, which is what this projection exists to avoid.

Also drops entries whose author is not the account whose wallet is
open, so a node answering with anything else cannot move the total,
and moves the builder to api/queries where app-specific queries live.
