+++ apps/web/src/api/queries/index.ts
@@ -3,3 +3,4 @@ export * from "./get-account-posts-feed-query";
 export * from "./get-gifs-query";
 export * from "./useClientTheme";
 export * from "./useHydrated";
+export * from "./pending-payouts-query";
